cdk/requirements.txt
aws-cdk-lib==2.177.0